requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
httpx
pytest
python-dotenv
boto3
google-generativeai
